requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
diff-match-patch==20230430
Django==5.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-import-export==4.1.1
django-print-settings==0.1.2
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
